WIP Add limit for nested data structures
YAML is not directly vulnerable to the Billion Laughs Attack:
https://en.wikipedia.org/wiki/Billion_laughs_attack#Variations

As long as the constructor uses perl references, everything is fine.

But if you print such data (for example as JSON), then it will get
huge, as JSON does not know references.

This commit is adding an experimental option that limits nested data.
Every anchor that is created gets assigned a number (its depth):
    &anchor [x] # 1
    &anchor [[x]] # 2

Whenever an alias is used, it will increase the total sum by its
depth.
Whenever an alias is used inside of another anchor, the depth of
this anchor will be increased by the alias depth.

    - &a [[x]]       # *a = 2
    - &b [*a,*a,*a]  # *b = 2+2+2 + 1 = 7  total=6
    - &c [*b]        # *c = 7 + 1          total=6+7=13

The total sum of depths must not be higher then a maximum which is
currently 256 by default.